move prepare-props into separate package
